requirements.txt
Django>=5.0
djangorestframework
python-dotenv
uvicorn
psycopg2-binary
concurrent-log-handler
selenium
geopandas>=0.14.0  # Replace GDAL with geopandas
djangorestframework-gis
geopandas
requests
tqdm
matplotlib